Refactor package structure: consolidate `cardsummary` into `activities/cardsummary`, update imports, and clean up redundant nullable checks.
