+++ src/reloading-remote/client/keyple-mobile-android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/ui/activities/cardsummary/TitlesRecyclerAdapter.kt
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: BSD-3-Clause
  ****************************************************************************** */
-package org.calypsonet.keyple.demo.reload.remote.ui.cardsummary
+package org.calypsonet.keyple.demo.reload.remote.ui.activities.cardsummary
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
